# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface AuthUser {
  id: number;
  name: string;
  email: string;
}

export interface SignUpInput {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @minLength 8 */
  password: string;
}

export interface SignInInput {
  email: string;
  /** @minLength 1 */
  password: string;
}

export interface AuthResponse {
  user: AuthUser;
}

export type TrackerPageAutoCheck = typeof TrackerPageAutoCheck[keyof typeof TrackerPageAutoCheck];


export const TrackerPageAutoCheck = {
  none: 'none',
  tick: 'tick',
  cross: 'cross',
} as const;

export type TrackerPageTimer = typeof TrackerPageTimer[keyof typeof TrackerPageTimer];


export const TrackerPageTimer = {
  never: 'never',
  '1h': '1h',
  '6h': '6h',
  '12h': '12h',
  '24h': '24h',
} as const;

export interface TrackerCycle {
  id: string;
  habitName: string;
  period: string;
  checked: number;
  crossed: number;
  total: number;
  savedAt: string;
}

export type TrackerPageTrackerData = {[key: string]: 'none' | 'checked' | 'crossed'};

export interface TrackerPage {
  id: string;
  habitName: string;
  color: string;
  icon: string;
  startDate: string;
  daysCount: number;
  autoCheck: TrackerPageAutoCheck;
  timer: TrackerPageTimer;
  trackerData: TrackerPageTrackerData;
  history: TrackerCycle[];
}

export interface TrackerState {
  profileName: string;
  pages: TrackerPage[];
  activePageId: string;
}

export interface InsightDay {
  day: string;
  checked: number;
  crossed: number;
}

export interface TrackerInsights {
  totalPages: number;
  totalChecked: number;
  totalCrossed: number;
  bestStreak: number;
  completionRate: number;
  activeDays: number;
  weekly: InsightDay[];
}

export interface ReportInput {
  /** @minLength 5 */
  message: string;
  email?: string;
  name?: string;
}

export interface ReportResult {
  sent: boolean;
  message: string;
}