# lib/api-zod/src/generated/types/reportInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ReportInput {
  /** @minLength 5 */
  message: string;
  email?: string;
  name?: string;
}