# lib/api-zod/src/generated/types/signUpInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SignUpInput {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @minLength 8 */
  password: string;
}