# lib/api-zod/src/generated/types/trackerCycle.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TrackerCycle {
  id: string;
  habitName: string;
  period: string;
  checked: number;
  crossed: number;
  total: number;
  savedAt: Date;
}