# lib/api-zod/src/generated/types/trackerInsights.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { InsightDay } from './insightDay';

export interface TrackerInsights {
  totalPages: number;
  totalChecked: number;
  totalCrossed: number;
  bestStreak: number;
  completionRate: number;
  activeDays: number;
  weekly: InsightDay[];
}