# lib/api-zod/src/generated/types/trackerPage.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TrackerCycle } from './trackerCycle';
import type { TrackerPageAutoCheck } from './trackerPageAutoCheck';
import type { TrackerPageTimer } from './trackerPageTimer';
import type { TrackerPageTrackerData } from './trackerPageTrackerData';

export interface TrackerPage {
  id: string;
  habitName: string;
  color: string;
  icon: string;
  startDate: Date;
  daysCount: number;
  autoCheck: TrackerPageAutoCheck;
  timer: TrackerPageTimer;
  trackerData: TrackerPageTrackerData;
  history: TrackerCycle[];
}