# lib/api-zod/src/generated/types/trackerState.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TrackerPage } from './trackerPage';

export interface TrackerState {
  profileName: string;
  pages: TrackerPage[];
  activePageId: string;
}